requirements.txt
Flask
cryptography
google-cloud-storage
google-cloud-kms
google-auth
google-auth-oauthlib
pyjwt[crypto]
python-dateutil
